% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiralophtha_adga}
\alias{admiralophtha_adga}
\title{Geographic Atrophy Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11508 rows and 103 columns.
}
\source{
Derived from the OE and ADSL datasets using \code{{admiral}}, \code{{admiralophtha}} and the
\href{https://github.com/pharmaverse/admiralophtha/blob/main/inst/templates/ad_adga.R}{ADGA template}.
}
\usage{
admiralophtha_adga
}
\description{
An example Geographic Atrophy analysis dataset
}
\seealso{
Other datasets:
\code{\link{admiralophtha_adbcva}},
\code{\link{admiralophtha_adoe}},
\code{\link{admiralophtha_advfq}}
}
\concept{datasets}
\keyword{datasets}
