% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiralophtha_adoe}
\alias{admiralophtha_adoe}
\title{Ophthalmology Exam Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19136 rows and 103 columns.
}
\source{
Derived from the OE and ADSL datasets using \code{{admiral}}, \code{{admiralophtha}} and the
\href{https://github.com/pharmaverse/admiralophtha/blob/main/inst/templates/ad_adoe.R}{ADOE template}.
}
\usage{
admiralophtha_adoe
}
\description{
An example Ophthalmology Exam Analysis dataset
}
\seealso{
Other datasets:
\code{\link{admiralophtha_adbcva}},
\code{\link{admiralophtha_adga}},
\code{\link{admiralophtha_advfq}}
}
\concept{datasets}
\keyword{datasets}
