% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiralophtha_advfq}
\alias{admiralophtha_advfq}
\title{Visual Function Questionnaire Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 972 rows and 93 columns.
}
\source{
Derived from the ADSL and QS datasets using \code{{admiral}}, \code{{admiralophtha}}  and the
\href{https://github.com/pharmaverse/admiralophtha/blob/main/inst/templates/ad_advfq.R}{ADVFQ template}.
The full, open-source VFQ questionnaire can be accessed
\href{https://www.nei.nih.gov/about/education-and-outreach/outreach-materials/visual-function-questionnaire-25}{here}.
}
\usage{
admiralophtha_advfq
}
\description{
An example Visual Function Questionnaire (VFQ) analysis dataset
}
\seealso{
Other datasets:
\code{\link{admiralophtha_adbcva}},
\code{\link{admiralophtha_adga}},
\code{\link{admiralophtha_adoe}}
}
\concept{datasets}
\keyword{datasets}
