% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiralophtha_ex}
\alias{admiralophtha_ex}
\title{Exposure Ophthalmology SDTM Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 591 rows and 19 columns.
}
\source{
Constructed using \code{ex} from the \code{{admiral}} package
}
\usage{
admiralophtha_ex
}
\description{
An example Exposure SDTM dataset with ophthalmology-specific variables such as EXLOC and EXLAT
}
\seealso{
Other datasets: 
\code{\link{admiralophtha_adbcva}},
\code{\link{admiralophtha_adoe}},
\code{\link{admiralophtha_advfq}},
\code{\link{admiralophtha_qs}}
}
\concept{datasets}
\keyword{datasets}
