% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiralophtha_qs}
\alias{admiralophtha_qs}
\title{Ophthalmology Questionnaires SDTM Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 197671 rows and 20 columns.
}
\source{
Constructed using \code{qs} from the \code{{admiral}} package
}
\usage{
admiralophtha_qs
}
\description{
An example Questionnaires SDTM dataset with ophthalmology-specific questionnaire of NEI VFQ-25
}
\seealso{
Other datasets: 
\code{\link{admiralophtha_adbcva}},
\code{\link{admiralophtha_adoe}},
\code{\link{admiralophtha_advfq}},
\code{\link{admiralophtha_ex}}
}
\concept{datasets}
\keyword{datasets}
