% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{advfq}
\alias{advfq}
\title{Visual Function Questionnaire Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 0 rows and 85 columns.
}
\source{
Derived from the \code{ADSL} and \code{qs} datasets using \verb{\{admiral}, \code{{admiralophtha}} and
(\url{https://github.com/pharmaverse/admiralophtha/blob/main/inst/templates/ad_advfq.R})
}
\usage{
advfq
}
\description{
An example Visual Function Questionnaire (VFQ) analysis dataset
}
\seealso{
Other datasets: 
\code{\link{adbcva}},
\code{\link{adoe}}
}
\concept{datasets}
\keyword{datasets}
