% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_bcvacritxfl.R
\name{derive_var_bcvacritxfl}
\alias{derive_var_bcvacritxfl}
\title{Adds \code{CRITx}/\code{CRITxFL} pairs to BCVA dataset}
\usage{
derive_var_bcvacritxfl(
  dataset,
  crit_var,
  bcva_ranges = NULL,
  bcva_uplims = NULL,
  bcva_lowlims = NULL,
  additional_text = "",
  critxfl_index = NULL
)
}
\arguments{
\item{dataset}{Input dataset containing BCVA data (usually ADBCVA).

\describe{
\item{Permitted values}{a dataset, i.e., a \code{data.frame} or tibble}
\item{Default value}{none}
}}

\item{crit_var}{Variable with respect to which \code{CRITx}/\code{CRITxFL} are derived
(usually \code{CHG} or \code{AVAL}).

\describe{
\item{Permitted values}{an unquoted symbol, e.g., \code{AVAL}}
\item{Default value}{none}
}}

\item{bcva_ranges}{List of numeric vectors. For each
vector \code{c(a,b)} in \code{bcva_ranges}, a pair of variables \code{CRITx}, \code{CRITxFL} is created
with the condition: \verb{a <=  crit_var <= b}. If criterion flags of that type are not
required, then leave as \code{NULL}.

\describe{
\item{Permitted values}{a list containing one or more numeric vectors, each of length two. E.g.
\verb{list(c(1, 2), c(3, 4)}}
\item{Default value}{\code{NULL}}
}}

\item{bcva_uplims}{List containing one or more numeric elements. For each
element a in \code{bcva_uplims}, a pair of variables \code{CRITx}, \code{CRITxFL} is created
with the condition: \code{crit_var <= a}. If criterion flags of that type are not
required, then leave as \code{NULL}.

\describe{
\item{Permitted values}{a list containing one or more numeric scalars. E.g. \code{list(2, -4)}}
\item{Default value}{\code{NULL}}
}}

\item{bcva_lowlims}{List containing one or more numeric elements. For each
element b in \code{bcva_lowlims}, a pair of variables \code{CRITx}, \code{CRITxFL} is created
with the condition: \code{crit_var >= b}. If criterion flags of that type are not required,
then leave as \code{NULL}.

\describe{
\item{Permitted values}{a list containing one or more numeric scalars. E.g. \code{list(2, -4)}}
\item{Default value}{\code{NULL}}
}}

\item{additional_text}{string containing additional text to append to \code{CRITx}.

\describe{
\item{Permitted values}{a character scalar, i.e., a character vector of length one}
\item{Default value}{\code{""}}
}}

\item{critxfl_index}{positive integer detailing the first value of x to use
in \code{CRITxFL}. If not supplied, the function takes the first available value of
x, counting up from x = 1.

\describe{
\item{Permitted values}{a positive integer, e.g. \code{2} or \code{5}}
\item{Default value}{\code{NULL}}
}}
}
\value{
The input BCVA dataset with additional column pairs\code{CRITx}, \code{CRITxFL}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
The \code{derive_var_bcvacritxfl()}
function has been deprecated in favor of \code{admiral::derive_vars_crit_flag()} - please see
the \href{https://pharmaverse.github.io/admiralophtha/articles/adbcva.html#critflags}{criterion flag section of the ADBCVA vignette}
for more details.

Adds a criterion variables \code{CRITx} and their corresponding flags \code{CRITxFL} to a
dataset containing BCVA records
}
\details{
This function works by calling \code{derive_var_bcvacritxfl()} once for each of the
elements in \code{bcva_ranges}, \code{bcva_uplims} and \code{bcva_lowlims}.
NOTE: if \code{crit_var} is equal to \code{NA}, then the resulting criterion flag is also marked
as \code{NA}.
}
\seealso{
Other deprecated:
\code{\link[=derive_var_bcvacritxfl_util]{derive_var_bcvacritxfl_util()}}
}
\author{
Edoardo Mancini
}
\concept{deprecated}
\keyword{deprecated}
\keyword{der_ophtha}
