% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etdrs_to_logmar.R
\name{etdrs_to_logmar}
\alias{etdrs_to_logmar}
\title{ETDRS --> LogMAR conversion}
\usage{
etdrs_to_logmar(variable)
}
\arguments{
\item{variable}{variable containing ETDRS score to convert to logMAR}
}
\value{
The input variable converted
}
\description{
Convert ETDRS --> LogMAR conversion in (\code{AVAL})
}
\details{
ETDRS is converted by logMAR = -0.02 * ETDRS + 1.7
}
\examples{
library(tibble)
library(dplyr)
library(admiral)

adbcva <- tribble(
  ~STUDYID, ~USUBJID, ~AVAL,
  "XXX001", "P01", 5,
  "XXX001", "P02", 10,
  "XXX001", "P03", 15,
  "XXX001", "P04", 20,
  "XXX001", "P05", 25
)

adbcva <- adbcva \%>\% mutate(AVAL = etdrs_to_logmar(AVAL))
}
\author{
Rachel Linacre
}
\keyword{der_adbcva}
\keyword{der_adoe}
