% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiralophtha-package.R
\docType{package}
\name{admiralophtha-package}
\alias{admiralophtha}
\alias{admiralophtha-package}
\title{admiralophtha: ADaM in R Asset Library - Ophthalmology}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Aids the programming of Clinical Data Standards Interchange Consortium (CDISC) compliant Ophthalmology Analysis Data Model (ADaM) datasets in R. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, \url{https://www.cdisc.org/standards/foundational/adam/adamig-v1-3-release-package}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiralophtha/}
  \item \url{https://github.com/pharmaverse/admiralophtha/}
  \item Report bugs at \url{https://github.com/pharmaverse/admiralophtha/issues/}
}

}
\author{
\strong{Maintainer}: Edoardo Mancini \email{edoardo.mancini@roche.com} (\href{https://orcid.org/0009-0006-4899-8641}{ORCID})

Authors:
\itemize{
  \item Edoardo Mancini \email{edoardo.mancini@roche.com} (\href{https://orcid.org/0009-0006-4899-8641}{ORCID})
  \item Yuki Matsunaga \email{yuki.matsunaga@novartis.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item Novartis AG [copyright holder, funder]
}

}
\keyword{internal}
