% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiralophtha-package.R
\docType{package}
\name{admiralophtha-package}
\alias{admiralophtha}
\alias{admiralophtha-package}
\title{admiralophtha: ADaM in R Asset Library - Ophthalmology}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Aids the programming of Clinical Data Standards Interchange Consortium (CDISC) compliant Ophthalmology Analysis Data Model (ADaM) datasets in R. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, \url{https://www.cdisc.org/standards/foundational/adam/adamig-v1-3-release-package}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiralophtha/}
  \item \url{https://github.com/pharmaverse/admiralophtha/}
}

}
\author{
\strong{Maintainer}: Edoardo Mancini \email{edoardo.mancini@roche.com}

Authors:
\itemize{
  \item Gordon Miller \email{gordon.miller@roche.com}
  \item Ritika Aggarwal \email{ritika.aggarwal@novartis.com}
  \item Jane Gao \email{jane.gao@roche.com}
  \item William Holmes \email{william.holmes@roche.com}
  \item Rachel Linacre \email{rachel.linacre@roche.com}
  \item Lucy Palmen \email{lucy.palmen@roche.com}
  \item Nandini R Thampi \email{nandini.rajendhiran@novartis.com}
  \item Yuki Matsunaga \email{yuki.matsunaga@novartis.com}
}

}
\keyword{internal}
