% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiralophtha_adoe}
\alias{admiralophtha_adoe}
\title{Ophthalmology Exam Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7672 rows and 98 columns.
}
\source{
Derived from the \code{oe} and \code{ADSL} datasets using \verb{\{admiral}, \code{{admiralophtha}} and
(\url{https://github.com/pharmaverse/admiralophtha/blob/main/inst/templates/ad_adoe.R})
}
\usage{
admiralophtha_adoe
}
\description{
An example Ophthalmology Exam Analysis dataset
}
\seealso{
Other datasets: 
\code{\link{admiralophtha_adbcva}},
\code{\link{admiralophtha_advfq}},
\code{\link{admiralophtha_ex}}
}
\concept{datasets}
\keyword{datasets}
