% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmar_etdrs.R
\name{convert_etdrs_to_logmar}
\alias{convert_etdrs_to_logmar}
\title{ETDRS --> LogMAR conversion}
\usage{
convert_etdrs_to_logmar(value)
}
\arguments{
\item{value}{object containing ETDRS score to convert to logMAR.

\describe{
\item{Permitted values}{a numeric value, e.g. \code{2}, \code{-5}, \code{1.4}}
\item{Default value}{none}
}}
}
\value{
The input value converted converted to logMAR units.
}
\description{
Convert ETDRS score to LogMAR units
}
\details{
ETDRS value converted to logMAR as:

\deqn{logMAR = -0.02 * ETDRS + 1.7}

Source for conversion formula:
Beck, R.W., et al. A computerized method of visual acuity testing.
American Journal of Ophthalmology, 135(2), pp.194-205.
doi:https://doi.org/10.1016/s0002-9394(02)01825-1.
}
\examples{
library(tibble)
library(dplyr)
library(admiral)

oe <- tribble(
  ~STUDYID, ~USUBJID, ~OETESTCD, ~OEMETHOD, ~OESTRESN,
  "XXX001", "P01", "VACSCORE", "logMAR EYE CHART", 1.08,
  "XXX001", "P02", "VACSCORE", "logMAR EYE CHART", 1.66,
  "XXX001", "P03", "VACSCORE", "logMAR EYE CHART", 1.60,
  "XXX001", "P04", "VACSCORE", "ETDRS EYE CHART", 57,
  "XXX001", "P05", "VACSCORE", "ETDRS EYE CHART", 62
)

adbcva <- oe \%>\%
  filter(OETESTCD == "VACSCORE" & toupper(OEMETHOD) == "ETDRS EYE CHART") \%>\%
  mutate(OESTRESN = convert_etdrs_to_logmar(OESTRESN))
}
\seealso{
\code{\link[=convert_logmar_to_etdrs]{convert_logmar_to_etdrs()}}
}
\author{
Rachel Linacre
}
\keyword{utils_fmt}
